Fix for JpegEncoder to rewrite DPI info
Jpeg files can store this information in JFIF and EXIF. The EXIF values
are kept to see the image metadata as it was shot. The Jpeg JFIF header
for APP0 will be modified if it exists or added if the Image was not a
JFIF. This library always adds JFIF headers to Jpeg files.