Fix for hierarchical Jpeg files
These kinds of files weren't supported because they have multiple Jpeg
frames. These frames are likely used by digital cameras to be able to
produce fast renders on limited hardware and bandwidth. The JFIF
Extension specification reports that the JPEG decoder will be
responsible for decoding the appropriate frame depending on the context.

This commit always decodes the biggest frame into the resulting image.
If developers want control of this feature a property on the decoder
that defines the behavior should suffice.